Bluetooth: TBS: Make GTBS mandatory
Any places that checked for GTBS is now always enabled
as GTBS is mandatory to support, and the Kconfig option
was removed.

Also removed a duplicate Kconfig option for TBS,
BT_TBS_SERVICE_COUNT, as it was unused.

The other Kconfig option BT_TBS_BEARER_COUNT is used
instead exclusively, and has been slightly prepared for
the option to be 0 by setting the value explicit for
tests and samples that need it.

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>